+++ src/Application/Request.ts
@@ -1,0 +1,4 @@
+/// <amd-module name="Application/Request" />
+import { default as Request } from "Application/_Request/Request";
+export default Request;
+export { default as Store } from "Application/_Request/Store";
